+++ sentry-android-ui-tests/end2end/src/androidTest/java/io/sentry/android/uitests/end2end/mockservers/RelayAsserter.kt
@@ -41,5 +41,5 @@ class RelayAsserter(
         }
     }
 
-    data class RelayResponse (val request: RecordedRequest, val response: MockResponse)
+    data class RelayResponse(val request: RecordedRequest, val response: MockResponse)
 }