+++ sentry-android-ui-tests/end2end/src/main/java/io/sentry/android/uitests/end2end/EmptyActivity.kt
@@ -2,7 +2,6 @@ package io.sentry.android.uitests.end2end
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import io.sentry.Sentry
 
 class EmptyActivity : AppCompatActivity() {