+++ sentry-android-ui-tests/end2end/src/main/java/io/sentry/android/uitests/end2end/utils/BooleanIdlingResource.kt
@@ -25,5 +25,4 @@ class BooleanIdlingResource(private val name: String) : IdlingResource {
     override fun registerIdleTransitionCallback(callback: IdlingResource.ResourceCallback?) {
         this.callback = callback
     }
-
 }